backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pydantic-settings==2.8.1
reportlab==4.4.1
